Mac-MCP/ControlPanel/Sources/Lotus/Resources/pyproject.toml
[project]
name = "mac-mcp"
version = "0.1.0"
description = "Mac-MCP – AI macOS Desktop Control via MCP"
authors = [
    { name = "Jayash Bhandary" }
]
readme = "README.md"
license = { file = "LICENSE.md" }
keywords = ["mac", "mcp", "ai", "desktop", "automation", "macos"]
requires-python = ">=3.13"
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3.13",
    "Topic :: Desktop Environment",
]
dependencies = [
    "click>=8.2.1",
    "customtkinter>=5.2.0",
    "fastmcp>=3.0",
    "pillow>=11.2.1",
    "mss>=9.0.0",
    "pynput>=1.7.0",
    "pyobjc-framework-ApplicationServices>=10.0",
    "pyobjc-framework-Quartz>=10.0",
    "pyobjc-framework-Cocoa>=10.0",
    "atomacos>=3.0.0",
    "psutil>=7.0.0",
    "fuzzywuzzy>=0.18.0",
    "python-levenshtein>=0.27.1",
    "markdownify>=1.1.0",
    "requests>=2.32.3",
    "posthog>=7.4.0",
    "platformdirs>=4.3.8",
    "python-dotenv>=1.0.0",
    "tabulate>=0.9.0",
    "thefuzz>=0.22.1",
    "uuid7>=0.1.0",
    "playwright>=1.42.0",
    "cryptography>=42.0.0",
    "openai>=1.0.0",
    "google-genai>=0.5.0",
    "python-telegram-bot>=21.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.2",
    "reportlab>=4.1.0",
    "pytesseract>=0.3.10",
    "ollama>=0.4.0",
    "yt-dlp>=2024.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
mac-mcp = "mac_mcp.__main__:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
